HR Analytics
Although the analysis and the charts look pretty simple but a way more detailed analysis is done in Tableau using level of detail (LOD). LOD are the expressions that allow you to perform mathematical operations at a certain level. You can perform the calculation / operation on the entire dataset and also on the group of a dataset that is created within the Tableau (Average Income of the entire company and average income for each age group in this analysis) and match them one against another in the same view. 

For more detailed explanation please give it a read, https://www.tableau.com/about/blog/LOD-expressions  and https://help.tableau.com/current/pro/desktop/en-us/calculations_calculatedfields_lod.htm.

Key findings: 

From the segmented bar chart (horizontal) that shows Employee count by age group, we can say that the majority number of people leaving the company lie in the age group of 26 - 35. 
Histogram for Employee count by Income shows that the people with monthly income between $2,000 - $4,000 are more likely to leave the organization. 
The positive - negative bar chart is indicating that the people between the age group of 18 - 25 are getting paid ($3000) way below the average monthly salary of the company ($7,000). This also justifies our 2nd observation but not completely agrees with the 1st point. 
The second segmented bar chart (vertical) depicts that the maximum employees leave job in their 2nd year with the company. 
We can conclude one thing from this dashboard is that the younger employees are more likely to leave the company and that too at the early phase of their association with the company. But we still cannot come up with the most appropriate reason that why the young people are more likely to leave the company. For that we might need more indicators such as employee’s daily commuting data (distance, address, etc.), hours the employee is logged in to the system, an employee’s hours of collaboration with the team members and some other parameters. 

Please suggest what all indicators can be included to make this analysis more useful for the HR of the company. 

To see my work please go to: https://public.tableau.com/profile/varun.ghorpade#!/vizhome/HRAnalytics_15953700031570/Dashboard1