Shipping Analytics
Shipping Analytics: 
Link to the viz: https://lnkd.in/eYniNRk

For any business to overcome the setbacks or any kind of shortcomings it is very important to first answer the 3 Ws, What, When and Where. To come up with an ultimate solution to any problem, we must start with the spot checks. This viz is also meant to answer such queries related shipping of the products.

In this I have created various calculated fields to find out the number of orders shipped or not shipped on a given date, total count of orders per category and several other KPIs.
I have made use of dual axis in Tableau Software to create the bar chart that answers whether or not the product in a particular is shipped or not in the selected month. dashboards